[ARM] add stride_slice_v2 layer

+++ source/tnn/layer/concat_layer.cc
@@ -35,7 +35,7 @@ inline bool ConcatLayerCheckShape(DimsVector shape1, DimsVector shape2, int excl
     }
 
     if (exclude_axis >= shape1.size()) {
-        LOGE("exclude_axis:%d out of shape size:%d\n", exclude_axis, shape1.size());
+        LOGE("exclude_axis:%d out of shape size:%d\n", exclude_axis, (int)shape1.size());
         return false;
     }
     return true;
